refactor: move fake stores
