+++ src/lib/features/feature-toggle/createFeatureToggleService.ts
@@ -18,8 +18,8 @@ import EnvironmentStore from '../../db/environment-store';
 import { Db } from '../../db/db';
 import { IUnleashConfig } from '../../types';
 import FakeEventStore from '../../../test/fixtures/fake-event-store';
-import FakeFeatureStrategiesStore from '../../../test/fixtures/fake-feature-strategies-store';
-import FakeFeatureToggleStore from '../../../test/fixtures/fake-feature-toggle-store';
+import FakeFeatureStrategiesStore from './fakes/fake-feature-strategies-store';
+import FakeFeatureToggleStore from './fakes/fake-feature-toggle-store';
 import FakeFeatureToggleClientStore from '../../../test/fixtures/fake-feature-toggle-client-store';
 import FakeProjectStore from '../../../test/fixtures/fake-project-store';
 import FakeFeatureEnvironmentStore from '../../../test/fixtures/fake-feature-environment-store';
