+++ src/lib/services/client-metrics/last-seen/createLastSeenService.ts
@@ -1,4 +1,4 @@
-import FakeFeatureToggleStore from '../../../../test/fixtures/fake-feature-toggle-store';
+import FakeFeatureToggleStore from '../../../features/feature-toggle/fakes/fake-feature-toggle-store';
 import FeatureToggleStore from '../../../features/feature-toggle/feature-toggle-store';
 import { Db, IUnleashConfig } from '../../../server-impl';
 import { FakeLastSeenStore } from './fake-last-seen-store';
