+++ src/test/fixtures/store.ts
@@ -1,7 +1,7 @@
-import FakeFeatureStrategiesStore from './fake-feature-strategies-store';
+import FakeFeatureStrategiesStore from '../../lib/features/feature-toggle/fakes/fake-feature-strategies-store';
 import FakeClientInstanceStore from './fake-client-instance-store';
 import FakeClientApplicationsStore from './fake-client-applications-store';
-import FakeFeatureToggleStore from './fake-feature-toggle-store';
+import FakeFeatureToggleStore from '../../lib/features/feature-toggle/fakes/fake-feature-toggle-store';
 import FakeTagStore from './fake-tag-store';
 import FakeTagTypeStore from './fake-tag-type-store';
 import FakeEventStore from './fake-event-store';
